+++ src/voxel/mod.rs
@@ -9,8 +9,7 @@ use std::time::Instant;
 
 use super::{
     fem::{
-        hex::{HexahedralFiniteElements, NUM_NODES_HEX},
-        Blocks, FiniteElements, NODE_NUMBERING_OFFSET,
+        Blocks, HexahedralFiniteElements, NODE_NUMBERING_OFFSET, HEX,
     },
     Connectivity, Coordinate, Coordinates, Octree, Tree, Vector, NSD,
 };
@@ -269,10 +268,10 @@ fn initial_element_node_connectivity(
     filtered_voxel_data: &VoxelDataSized<NSD>,
     nelxplus1: &usize,
     nelyplus1: &usize,
-) -> Connectivity<NUM_NODES_HEX> {
+) -> Connectivity<HEX> {
     #[cfg(feature = "profile")]
     let time = Instant::now();
-    let element_node_connectivity: Connectivity<NUM_NODES_HEX> = filtered_voxel_data
+    let element_node_connectivity: Connectivity<HEX> = filtered_voxel_data
         .iter()
         .map(|&[i, j, k]| {
             [
@@ -299,7 +298,7 @@ fn initial_element_node_connectivity(
 }
 
 fn initial_nodal_coordinates(
-    element_node_connectivity: &Connectivity<NUM_NODES_HEX>,
+    element_node_connectivity: &Connectivity<HEX>,
     filtered_voxel_data: &VoxelDataSized<NSD>,
     number_of_nodes_unfiltered: usize,
     scale: Scale,
@@ -366,7 +365,7 @@ fn initial_nodal_coordinates(
 }
 
 fn renumber_nodes(
-    element_node_connectivity: &mut Connectivity<NUM_NODES_HEX>,
+    element_node_connectivity: &mut Connectivity<HEX>,
     mut initial_nodal_coordinates: InitialNodalCoordinates,
     number_of_nodes_unfiltered: usize,
 ) -> Coordinates {
@@ -408,7 +407,7 @@ fn finite_element_data_from_data(
     remove: Option<Blocks>,
     scale: Scale,
     translate: &Vector,
-) -> Result<(Blocks, Connectivity<NUM_NODES_HEX>, Coordinates), String> {
+) -> Result<(Blocks, Connectivity<HEX>, Coordinates), String> {
     let shape = data.shape();
     let nelxplus1 = shape[0] + 1;
     let nelyplus1 = shape[1] + 1;
