+++ tests/fem.rs
@@ -1,4 +1,4 @@
-use automesh::{FiniteElements, Scale, Vector, Voxels, NSD};
+use automesh::{Scale, Vector, Voxels, NSD};
 use conspire::math::{Tensor, TensorArray, TensorRank1};
 use std::{
     fs::File,
